+++ profiles/foot.lua
@@ -6,6 +6,7 @@ Set = require('lib/set')
 Sequence = require('lib/sequence')
 Handlers = require("lib/way_handlers")
 find_access_tag = require("lib/access").find_access_tag
+country_speeds = require("lib/country_speeds")
 
 function setup()
   local walking_speed = 5
@@ -76,7 +77,7 @@ function setup()
       'motorroad'
     },
 
-    speeds = Sequence {
+    hwyspeeds = Sequence {
       highway = {
         primary         = walking_speed,
         primary_link    = walking_speed,
@@ -97,7 +98,9 @@ function setup()
         footway         = walking_speed,
         pier            = walking_speed,
       },
+    },
 
+    otherspeeds = Sequence {
       railway = {
         platform        = walking_speed
       },
@@ -115,6 +118,9 @@ function setup()
         track           = walking_speed
       }
     },
+
+    profile = 'foot',
+    uselocationtags = Set { },
 
     route_speeds = {
       ferry = 5
@@ -206,6 +212,13 @@ function process_way(profile, way, result)
     return
   end
 
+  local speedhandler = WayHandlers.speed
+
+  if profile.uselocationtags and profile.uselocationtags.countryspeeds then
+    data.location = country_speeds.getcountrytag(way)
+    speedhandler = country_speeds.wayspeed
+  end
+
   local handlers = Sequence {
     -- set the default mode for this profile. if can be changed later
     -- in case it turns we're e.g. on a ferry
@@ -231,7 +244,7 @@ function process_way(profile, way, result)
     WayHandlers.movables,
 
     -- compute speed taking into account way type, maxspeed tags, etc.
-    WayHandlers.speed,
+    speedhandler,
     WayHandlers.surface,
 
     -- handle turn lanes and road classification, used for guidance