+++ profiles/lib/way_handlers.lua
@@ -291,7 +291,26 @@ function WayHandlers.speed(profile,way,result,data)
     return        -- abort if already set, eg. by a route
   end
 
-  local key,value,speed = Tags.get_constant_by_key_value(way,profile.speeds)
+  local wayspeeds
+  if profile.speeds then
+    wayspeeds = profile.speeds
+  elseif profile.hwyspeeds then
+    wayspeeds = profile.hwyspeeds
+  else
+    return false
+  end
+
+  local key,value,speed = Tags.get_constant_by_key_value(way,wayspeeds)
+
+  if speed then
+    result.forward_speed = speed
+    result.backward_speed = speed
+  else
+    -- Check other speed categories (railway, amenity, man_made, leisure)
+    if profile.otherspeeds then
+      key,value,speed = Tags.get_constant_by_key_value(way,profile.otherspeeds)
+    end
+  end
 
   if speed then
     -- set speed by way type